Extracted out common webpack config to reduce duplication
