+++ package.json
@@ -7,8 +7,8 @@
     "dev": "npm-run-all --parallel build build-server redis",
     "dev:watch": "npm-run-all --parallel build:watch build-server:watch redis",
     "dev:serve": "npm-run-all --parallel dev:watch serve",
-    "build": "webpack --config webpack.config.js",
-    "build:watch": "webpack --config webpack.config.js --progress --colours --watch",
+    "build": "webpack --config webpack.browser.config.js",
+    "build:watch": "webpack --config webpack.browser.config.js --progress --colours --watch",
     "build-server": "webpack --config webpack.server.config.js",
     "build-server:watch": "webpack --config webpack.server.config.js --progress --colours --watch",
     "build:production": "NODE_ENV=production webpack -p --config webpack.production.config.js",
