+++ webpack.server.config.js
@@ -1,15 +1,42 @@
-require("env2")(__dirname + "/config.env");
 var path = require("path");
-var extractTextPlugin = require("extract-text-webpack-plugin");
 var webpack = require("webpack");
+var baseConfig = require("./webpack.base.config.js");
+
+require("env2")(__dirname + "/config.env");
 
 const BUILD_DIR = path.resolve(__dirname, "src/build/static");
 const SERVER_DIR = path.resolve(__dirname, "src/backend-hapi");
 
 process.env.RENDER_ENV = "server";
 
-module.exports = {
-    cache: true,
+var serverLoaders = [
+    {
+        test: /\.jsx?/,
+        include: __dirname + "/src",
+        exclude: /node_modules/,
+        loader: "babel",
+        query: {
+            cacheDirectory: true
+        }
+    },
+    {
+        test: /\.json$/,
+        loader: "json-loader"
+    }
+];
+
+var serverPlugins = [
+    new webpack.EnvironmentPlugin([
+        "NODE_ENV",
+        "RENDER_ENV",
+        "DB_URL",
+        "DB_MAX_CONNECTIONS",
+        "SK_KEY",
+        "SK_URL"]
+    )
+];
+
+var config = {
     target: "node",
     node: {
         __dirname: false
@@ -23,56 +50,12 @@ module.exports = {
         filename: "../server.bundle.js",
         libraryTarget: "commonjs2"
     },
-    externals: /^[a-z\-0-9]+$/,
-    devtool: "eval-source-map",
-    module: {
-        loaders: [
-            {
-                test: /\.json$/,
-                loader: "json-loader"
-            },
-            {
-                test: /\.jsx?/,
-                include: __dirname + "/src",
-                exclude: /node_modules/,
-                loader: "babel",
-                query: {
-                    cacheDirectory: true
-                }
-            },
-            {
-                test: /\.(css|scss|sass)$/,
-                exclude: /node_modules/,
-                loader: extractTextPlugin.extract("isomorphic-style-loader", "css!sass")
-            },
-            {
-                test: /\.(jpe?g|png|gif|svg)$/i,
-                loaders: [
-                    "url?limit=8192&name=images/[name].[ext]",
-                    "img?minimize&optimizationLevel=5&progressive=true"
-                ]
-            },
-            {
-                test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-                loader: "url-loader?limit=10000&mimetype=application/font-woff&name=fonts/[name].[ext]"
-            },
-            {
-                test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-                loader: "file-loader?name=fonts/[name].[ext]"
-            }
-        ]
-    },
-    plugins: [
-        new extractTextPlugin("styles/style.css", {
-            allChunks: true
-        }),
-        new webpack.EnvironmentPlugin([
-            "NODE_ENV",
-            "RENDER_ENV",
-            "DB_URL",
-            "DB_MAX_CONNECTIONS",
-            "SK_KEY",
-            "SK_URL"]
-        )
-    ]
+    externals: /^[a-z\-0-9]+$/
 };
+
+var serverConfig = Object.assign({}, baseConfig, config);
+serverConfig.module.loaders.push(...serverLoaders);
+serverConfig.plugins.push(...serverPlugins);
+
+module.exports = serverConfig;
+
