+++ webpack.test.config.babel.js
@@ -1,5 +1,6 @@
-import nodeExternals from 'webpack-node-externals';
+import nodeExternals from "webpack-node-externals";
 
-let config = require("./webpack.server.config.js");
+const config = require("./webpack.server.config.js");
 config.externals = [nodeExternals()];
+
 export default config;
